+++ djangosaml2idp/models/__init__.py
@@ -1,0 +1,5 @@
+from .ServiceProviderModel import ServiceProvider
+
+__all__ = [
+    "ServiceProvider"
+]
