+++ djangosaml2idp/mxins/LoginRequiredMixin.py
@@ -17,7 +17,8 @@ class LoginRequiredMixin(AccessMixin):
                 "spauthn", None) or request.COOKIES["spauthn"]
             token = Token.objects.get(key=spauthn) # pylint: disable=no-member
             if token:
+                request.user = token.user
                 return super().dispatch(request, tenant_uuid, app_id, *args, **kwargs)
         except Exception as err:    # pylint: disable=broad-except
             logger.debug(err)
-            return self.handle_no_permission(request.session['SAMLRequest'],tenant_uuid,app_id)
+            return self.handle_no_permission(request.session['SAMLRequest'],request.session["Binding"],tenant_uuid,app_id)
